все арифмитические действия верны сделала вывод целого числа в методе Print
